requirements.txt
aiosmtplib
annotated-types
anyio
bcrypt
blinker
boto3
botocore
certifi
click
colorama
contourpy
cycler
dnspython
ecdsa
email_validator
fastapi
fastapi-cli
fonttools
greenlet
h11
httpcore
httptools
httpx
hurry.filesize
idna
imageio
Jinja2
jmespath
kiwisolver
lazy_loader
markdown-it-py
MarkupSafe
matplotlib
mdurl
networkx
numpy
orjson
packaging
passlib
pillow
psycopg2-binary
pyasn1
pydantic
pydantic-settings
pydantic_core
Pygments
pyparsing
python-dateutil
python-dotenv
python-jose
python-multipart
PyYAML
rich
rsa
s3transfer
scikit-image
scipy
setuptools
six
sniffio
SQLAlchemy
starlette
tifffile
typer
typing_extensions
ujson
urllib3
uvicorn
watchfiles
websockets
alembic
scikit-learn
opencv-python